updated portal tests w/ new account, moved profile test to ignore
