+++ portal/tests/protractorConf.js
@@ -67,10 +67,10 @@ exports.config = {
   },
   params: {
     login: {
-      user: 'sfeldman+apijeep@apigee.com',
+      user: 'rsimpson+test@apigee.com',
       password: 'P@ssw0rd1'
     },
-    orgName:'sfeldmanapijeep',
+    orgName:'rodprotractortests',
     appName1:'website',
     appName2:'mobile',
     useSso:false
